fixe primary key repository type

+++ back/src/main/java/com/openclassrooms/mddapi/repository/ArticleRepository.java
@@ -3,5 +3,5 @@ package com.openclassrooms.mddapi.repository;
 import com.openclassrooms.mddapi.model.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<Article, Integer> {
+public interface ArticleRepository extends JpaRepository<Article, Long> {
 }
