+++ back/src/main/java/com/openclassrooms/mddapi/repository/ThemeRepository.java
@@ -3,5 +3,5 @@ package com.openclassrooms.mddapi.repository;
 import com.openclassrooms.mddapi.model.Theme;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ThemeRepository extends JpaRepository<Theme, Integer> {
+public interface ThemeRepository extends JpaRepository<Theme, Long> {
 }
